refactor(ui): move route definitions to a leaf module
Break the app-route-paths <-> tab-slugs import cycle by extracting the
static route table into app-route-definitions.ts and importing it from
both sides. Drop the assertion baseline entry that moved with the table.
