+++ ui/src/app-route-paths.ts
@@ -10,10 +10,19 @@ import {
 } from "@openclaw/uirouter";
 import type { RouteLocation } from "@openclaw/uirouter";
 import { isValidWorkboardBoardId } from "@openclaw/workboard-contract";
+import {
+  APP_ROUTE_DEFINITIONS,
+  APP_ROUTE_PATHS,
+  ROUTE_ID_BY_PATH,
+  type RouteId,
+} from "./app-route-definitions.ts";
 import { DEFAULT_AGENT_PANEL, isAgentsPanel, type AgentsPanel } from "./lib/agents/panels.ts";
 import type { BoardFace } from "./lib/board/settings.ts";
 import { takeGraphemes } from "./lib/graphemes.ts";
 import { pluginTabSlugFromPath } from "./pages/plugin/tab-slugs.ts";
+
+export { type RouteId };
+export { APP_ROUTE_IDS } from "./app-route-definitions.ts";
 export const INTERNAL_AGENT_PATH_PARAM = "__openclawAgentPath";
 export const INTERNAL_ACTIVITY_PATH_PARAM = "__openclawActivityPath";
 export const INTERNAL_SESSION_PATH_PARAM = "__openclawSessionPath";
@@ -34,84 +43,6 @@ type AgentRoutePath = {
   panelSegment: AgentsPanel | null;
   invalidPanel: boolean;
 };
-
-const APP_ROUTE_DEFINITIONS = {
-  settings: { path: "/settings" },
-  chat: { path: "/chat" },
-  dashboard: { path: "/dashboard" },
-  dashboards: { path: "/dashboards" },
-  custodian: { path: "/custodian" },
-  "new-session": { path: "/new" },
-  activity: { path: "/activity" },
-  meetings: { path: "/meetings" },
-  apps: { path: "/apps" },
-  portals: { path: "/portals" },
-  agents: { path: "/settings/agents", aliases: ["/agents"] },
-  channels: { path: "/settings/channels", aliases: ["/channels"] },
-  connection: { path: "/settings/connection" },
-  config: { path: "/settings/general", aliases: ["/config"] },
-  profile: { path: "/settings/profile", aliases: ["/profile"] },
-  communications: { path: "/settings/communications", aliases: ["/communications"] },
-  appearance: { path: "/settings/appearance", aliases: ["/appearance"] },
-  lobsterdex: { path: "/settings/lobsterdex", aliases: ["/lobsterdex"] },
-  device: { path: "/settings/device" },
-  "device-permissions": { path: "/settings/device/permissions" },
-  notifications: { path: "/settings/notifications" },
-  security: { path: "/settings/security" },
-  secrets: { path: "/settings/secrets" },
-  advanced: { path: "/settings/advanced" },
-  approvals: { path: "/settings/approvals" },
-  automation: { path: "/settings/automation", aliases: ["/automation"] },
-  mcp: { path: "/settings/mcp", aliases: ["/mcp"] },
-  memory: { path: "/settings/memory" },
-  talk: { path: "/settings/talk" },
-  infrastructure: { path: "/settings/infrastructure", aliases: ["/infrastructure"] },
-  labs: { path: "/settings/labs" },
-  updates: { path: "/settings/updates" },
-  about: { path: "/settings/about" },
-  "ai-agents": { path: "/settings/ai-agents", aliases: ["/ai-agents"] },
-  "model-setup": { path: "/settings/model-setup", aliases: ["/model-setup"] },
-  "model-providers": { path: "/settings/model-providers", aliases: ["/model-providers"] },
-  // Memory import, sessions, and worktrees are workspace destinations; the
-  // /settings/* aliases keep pre-restructure bookmarks and deep links working.
-  "memory-import": { path: "/memory-import", aliases: ["/settings/memory-import"] },
-  workboard: { path: "/workboard" },
-  worktrees: { path: "/worktrees", aliases: ["/settings/worktrees"] },
-  sessions: { path: "/sessions", aliases: ["/settings/sessions"] },
-  usage: { path: "/usage" },
-  debug: { path: "/debug" },
-  logs: { path: "/logs" },
-  "skill-workshop": { path: "/skills/workshop" },
-  skills: { path: "/skills" },
-  plugins: { path: "/settings/plugins" },
-  // Automations is the product name; /cron stays as a legacy alias for
-  // pre-rename bookmarks and deep links.
-  cron: { path: "/automations", aliases: ["/cron"] },
-  tasks: { path: "/tasks" },
-  devices: { path: "/settings/devices", aliases: ["/nodes"] },
-  "cloud-workers": { path: "/settings/cloud-workers" },
-  plugin: { path: "/plugin" },
-} as const;
-
-export type RouteId = keyof typeof APP_ROUTE_DEFINITIONS;
-export const APP_ROUTE_IDS = Object.keys(APP_ROUTE_DEFINITIONS) as RouteId[];
-const APP_ROUTE_PATHS: string[] = [];
-const ROUTE_ID_BY_PATH = new Map<string, RouteId>();
-// Static paths and aliases share one prepared index; earlier declarations keep priority.
-for (const routeId of APP_ROUTE_IDS) {
-  const definition = APP_ROUTE_DEFINITIONS[routeId];
-  const paths: readonly string[] =
-    "aliases" in definition ? [definition.path, ...definition.aliases] : [definition.path];
-  for (const path of paths) {
-    const normalizedPath = normalizePath(path);
-    APP_ROUTE_PATHS.push(normalizedPath);
-    if (!ROUTE_ID_BY_PATH.has(normalizedPath)) {
-      ROUTE_ID_BY_PATH.set(normalizedPath, routeId);
-    }
-  }
-}
-
-export const NATIVE_ROUTE_SEGMENTS = new Set(APP_ROUTE_PATHS.map((path) => path.split("/")[1]));
 
 export function isRouteId(routeId: string): routeId is RouteId {
   return Object.hasOwn(APP_ROUTE_DEFINITIONS, routeId);
